refactor: update global styles and HTML
Refactored global styles, setting 'Roboto' as the default font for body, input, text area, and buttons. Styles were also modified on the Header component where a color from the theme was applied to the 'span' element. In the HTML file, link tags were added to preconnect to the Google Fonts website and import the 'Baloo 2' and 'Roboto' font families.