feat: add SVG background to home page
An SVG background image was added to the Home page styles. The Intro component in the Home page styles was also modified to match the new design requirements. Changes in the Home component were made to reflect these style changes, basically wrapping the pre-existing Intro section with a Hero section incorporating the new background.