feat: add coffeeReducer and addNewCoffeeAction to reducers
Introduced a new reducer "coffeeReducer" to manage the state of coffee items in the cart. Also added an action "addNewCoffeeAction" that dispatches the type "ADD_COFFEE" with the new coffee item to be added to the state. This helps in managing the number of coffee items in the cart and updating quantities if the item already exists.